+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 Unreleased
-* Add support for loose validation
+* Add validators for uuid format (`Uuid.isValidUUIDFormat(...)`, `UuidValue.validateFormat()`, `UuidValidation.isValidUUIDFormat(...)` and `UuidValidation.isValidFormatOrThrow(...)`)
+* Add `UuidValue.withFormatValidation(...)` constructor
+* Add `Uuid.parseHex128(...)`, `Uuid.parseHex128AsByteList(...)`, `UuidParsing.parseHex128(...)` and `UuidParsing.parseHex128AsByteList(...)` methods
 
 v4.5.3